Moved bill logic to bills.py
